pyproject.toml
[project]
name = "build-your-own-deep-research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "exa-py>=2.7.0",
    "google-genai>=1.66.0",
    "logfire[google-genai]>=4.27.0",
    "prompt-toolkit>=3.0.52",
    "pydantic-ai-slim[google,logfire]>=1.67.0",
    "rich>=14.3.3",
    "textual>=8.1.1",
]
